+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## [Unreleased]
 
+## [0.0.3] - 2026-02-19
+
 ### Added
 
 - Automatic exclusion of Mix-generated and dependency directories (`_build`, `deps`, `.elixir_ls`, `.expert`) for
@@ -35,8 +37,7 @@ A new project has seen the light of day!
 - Initial scaffold created from
   [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
 
-[Unreleased]: https://github.com/mkaput/elixirij/compare/v0.0.2...HEAD
-
+[Unreleased]: https://github.com/mkaput/elixirij/compare/v0.0.3...HEAD
+[0.0.3]: https://github.com/mkaput/elixirij/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/mkaput/elixirij/compare/v0.0.1...v0.0.2
-
 [0.0.1]: https://github.com/mkaput/elixirij/commits/v0.0.1